data integrity check added
