+++ results/Detection_Results.json
@@ -3,7 +3,7 @@
     "First_name": "VOHAN ISRAEL",
     "Gender": "M",
     "Id_number": "080594100713",
-    "MRZ_l1": "IDFRAZIBI<<<<<<<<<<<<<<<<<<<<<9L1033",
-    "MRZ_l2": "0805941007132YOHAN<<ISRAEL<8104041M3",
+    "Mrz1": "IDFRAZIBI<<<<<<<<<<<<<<<<<<<<<941033",
+    "Mrz2": "0805941007132YOHAN<<ISRAEL<8104041M3",
     "Name": "ZIBI"
 }
